Added building position to cursor position
Added very basic building UI buttons. To add more buttons, just need to add more buildings to the buildings array in the BuildingManager.

Added Rigidbody and MeshCollider components to the 6 meshes that make up planets.

Added raycasting to move building objects to the point where the ray from the cursor hits the planet's new mesh colliders. The current building is then moved to that position.